Extract server_utils.py
